Removed named exports from Vue files
