+++ .vscode/settings.json
@@ -16,25 +16,81 @@
   "javascript.preferences.importModuleSpecifier": "non-relative",
   "typescript.preferences.importModuleSpecifier": "non-relative",
   "cSpell.words": [
+    "alertmanager",
+    "ALERTMANAGER",
+    "alertmanagerconfig",
+    "ALERTMANAGERCONFIG",
+    "apimachinery",
+    "apiregistration",
     "apiserver",
+    "apiservice",
+    "authconfig",
+    "authprovider",
     "autoscroll",
     "banzaicloud",
+    "bundledeployment",
+    "bundlenamespacemapping",
     "cacerts",
     "CAPI",
+    "catalogtemplate",
     "chainable",
+    "cloudcredential",
+    "clusterflow",
+    "clustergroup",
+    "clusteroutput",
+    "clusterregistrationtoken",
+    "clusterrepo",
+    "clusterrole",
+    "clusterrolebinding",
+    "clusterroletemplatebinding",
     "clusterroletemplatebindings",
+    "clustertemplate",
+    "clustertemplaterevision",
+    "cncf",
     "Codecov",
+    "configmap",
+    "constrainttemplate",
     "createapp",
+    "csidriver",
+    "daemonset",
+    "dedot",
+    "DEDOT",
+    "destinationrule",
+    "DETECTEXCEPTIONS",
+    "engineimage",
     "epinio",
+    "etcdbackup",
+    "etcdsnapshot",
+    "fleetworkspace",
     "fleetworkspaces",
+    "geoip",
+    "GEOIP",
+    "gitrepo",
+    "gitreporestriction",
+    "globaldnsprovider",
+    "globalrole",
+    "globalrolebinding",
     "hevi",
+    "horizontalpodautoscaler",
+    "httpconfig",
+    "ingressclass",
     "Istio",
+    "kontainer",
+    "KONTAINER",
+    "kontainerdriver",
     "kube",
     "kubeconfig",
     "kubectl",
     "Kubernetes",
     "kubevirt",
+    "limitrange",
+    "machinedeployment",
+    "machineset",
+    "managedchart",
     "mgmt",
+    "multiclusterapp",
+    "networkattachmentdefinition",
+    "networkpolicy",
     "neuvector",
     "nodedriver",
     "nodemetrics",
@@ -43,19 +99,49 @@
     "opsgenie",
     "OPSGENIE",
     "overcommit",
+    "pagerduty",
+    "PAGERDUTY",
+    "persistentvolume",
+    "persistentvolumeclaim",
+    "poddisruptionbudget",
+    "podmetrics",
+    "podmonitor",
+    "PODMONITOR",
+    "podsecurityadmissionconfigurationtemplate",
     "prefs",
     "prepending",
+    "projecthelmchart",
+    "PROJECTHELMCHART",
+    "projectroletemplatebinding",
     "projectsnamespaces",
+    "prometheusrule",
+    "PROMETHEUSRULE",
     "protip",
     "PSPS",
     "PVC",
     "pvcs",
+    "rancherusernotification",
+    "replicaset",
+    "replicationcontroller",
+    "resourcequota",
+    "rolebinding",
+    "roletemplate",
+    "serviceaccount",
+    "servicemonitor",
+    "SERVICEMONITOR",
     "shortkey",
+    "statefulset",
     "storageclass",
+    "SUMOLOGIC",
     "testid",
+    "thanosruler",
+    "THANOSRULER",
     "tolerations",
+    "uiplugin",
+    "userpreference",
     "userpreferences",
     "virtualmachine",
+    "virtualservice",
     "vuex",
     "whatsnew"
   ],
