+++ shell/plugins/formatters.js
@@ -3,7 +3,7 @@
 // We do it here to keep it away from plugins
 // It was in SortableTable itself, but this causes plugins to pull in all formatters and their dependencies
 
-import { FORMATTERS } from '@shell/components/SortableTable';
+import { FORMATTERS } from '@shell/components/SortableTable/sortable-config';
 
 const components = require.context('@shell/components/formatter', false, /[A-Z]\w+\.(vue)$/);
